add support for svn changelists
svn cangelists are supported here by adding command-line processing,
eg: --cl=foo, to add the given changelist to the list of files to add.
This is implemented by running svn st --cl for the passed-in argument.

To get this working, some of the code is moved around a little, notably
using a files=() array throughout to keep track of the list of passed-in
and --cl specified arguments.

As part of the argument processing, unknown arguments are detected, and
-- is taken to mean "interpret all further arguments as files". A new
fully_qualified() function is used in place of 'readlink -fn' to ensure
leading dashes are handled properly.

Multiline filenames output by 'svn st' have not been tested, and it is
assumed that they would break this code.

As command-line arguments were added here, a usage message was also
added, triggered by 'svn-cip help' (for consistency with svn), or
'svn-cip --help'