rewrite prune-remote as prune-merged
this script still makes some assumptions as to the structure of one's
branch names, but hopefully makes fewer assumptions / more-universal
ones.

prune-merged is intended to help "clean up" completed branches. A branch
is considered to be "complete" when it is merged into the "mainline",
mainline defaulting to refs/remotes/<remote>/master